+++ RELEASES.md
@@ -5,11 +5,22 @@ Ibis AdapterFramework release notes
 [JavaDocs](https://javadoc.ibissource.org/latest/)
 
 
-7.2-B1
+
+7.3-B1
 --------
 
 [Commits](https://github.com/ibissource/iaf/compare/v7.1...HEAD)
 [![Build Status](https://travis-ci.org/ibissource/iaf.png)](https://travis-ci.org/ibissource/iaf)
+
+- Generate IbisDoc and XSD and support beautiful configuration xml. The XSD can be used for code completion of beautiful Ibis configurations in Eclipse
+
+
+
+7.2
+--------
+
+[Commits](https://github.com/ibissource/iaf/compare/v7.1-B4...v7.1)
+[![Build Status](https://travis-ci.org/ibissource/iaf.png?branch=v7.1)](https://travis-ci.org/ibissource/iaf)
 
 - Fix NPE in BatchFileTransformerPipe when using an IbisLocalSender
 - Various bugfixes en performance improvements in SOAPProviders (WebServiceListener)
@@ -26,6 +37,7 @@ Ibis AdapterFramework release notes
     - Fix ClassCastException in Test tab for reports with description
     - Prevent losing typed data in edit mode on close or select node in tree
     - Copy report name from original report on Replace
+
 
 
 7.1